data: Move file review freshness checks
File-review state persists the last reviewed file and validates follow-up
action scope for page-aware command flows.

Freshness matching also lived in that module, so status reporting and
validation code imported persistence state for selected-file and snapshot
comparison.

This commit addresses that by moving selected review freshness checks into
data.file_review_freshness, routing status through that boundary, and
keeping validation imports private in file_review_state.

The next commit will add architecture coverage for the freshness boundary.
